文章卡片、个人信息、评论

+++ src/main/java/org/example/quantumblog/util/Generator.java
@@ -17,7 +17,7 @@ public class Generator {
     }
 
     //ID生成器
-    public long generateUniqueNumber(String username, long timestamp) {
+    public int generateUniqueNumber(String username, long timestamp) {
         Random random = new Random(username.hashCode() + timestamp);
         int number;
         do {
